% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements-data.R
\docType{data}
\name{elements}
\alias{elements}
\title{Elements table for EDS analysis}
\format{A data.frame containing info from the periodic table, obtained from package PeriodicTable}
\source{
\href{https://cran.r-project.org/web/packages/PeriodicTable/PeriodicTable.pdf}{PeriodicTable}
}
\usage{
data(elements)
}
\description{
Data frame containing information of the periodic table of elements.
The data frame is based on the package PeriodicTable, data(periodicTable). The only addition to this data frame is the last two columns (EDS and EDS2) containing the energy (KeV) table for EDS analysis
}
\examples{
data(elements)
}
\keyword{elements}
