% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ttaov.R
\name{Tt.aov}
\alias{Tt.aov}
\title{Direct estimation of 1 or 2 way anova}
\usage{
Tt.aov(vx, vtreat1, vtreat2)
}
\arguments{
\item{vx}{dependent variable}

\item{vtreat}{independent variable(s). If vtreat2 exist, the function will perform a 2-way anova}
}
\value{
aov.output a list with the following objects
\enumerate{
\item anova : Which contains the details of the analysis
\item stats : mean, sd and size of sample (n)
\item p : value of p from anova, for a quick answer
\item Tukey : If p>0.05, tukeyHSD is performed
}
}
\description{
Performs the analysis of variance using the functions aov and anova. Additionally, creates a table with mean values, standard deviation and size of sample per treatment. If p>0.05, a Tukey test is also performed.
}
\keyword{TukeyHSD}
\keyword{anova}
