% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBarplot.R
\name{Tt.Barplot}
\alias{Tt.Barplot}
\title{Bar chart with standard error lines using ggplot}
\usage{
Tt.Barplot(aov.output, treat2 = FALSE)
}
\arguments{
\item{aov.output}{is a data frame containing the statistics information with the following as column names.
\describe{
\item{Treatment}{The vector with the independent variables}
\item{x.mean}{Vector with mean values}
\item{x.sd}{vector with standard deviation}
\item{x.n}{vectore with the sample sizes}
}
The object \emph{stats} obtained from the function \code{Tt.aov()} can be used directly, as it has the specified format.}

\item{treat2=FALSE}{plots 1-way anova. If a second independent variable exists, change to TRUE}
}
\value{
gBarPlo An object for ggplot graphics
}
\description{
Bar chart with standard error lines using ggplot
}
\examples{
plot <- Tt.ggBarplot(data$stats) # loading the plot directly from stats object
plot + labs(title="Graphic Title", subtitle='Subtitle', x='labels in X', y='labels in Y') + scale_fill_brewer(palette=3, name='Treatments') # ggplot plotting
}
\keyword{anova}
\keyword{barplot}
\keyword{ggplot}
