% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ttreadabsobance.R
\name{Tt.read.abs}
\alias{Tt.read.abs}
\title{Calculation of a linear model to read and interprete absorbance}
\usage{
Tt.read.abs(model.concentrations, model.absorbance, predict = NULL,
  show.plot = TRUE)
}
\arguments{
\item{model.concentrations}{a vector containing the known concentrations of the standards (the order must match with that of absorbance)}

\item{model.absorbance}{a vector containing the absorbance read on the standards (the order must match with that of concentrations)}

\item{predict}{a vector containing absorbance, in order to predict concentrations}

\item{show.plot}{if TRUE, the data is plotted showing the model}
}
\value{
The function returns a list containing the details of the model and the predicted values if added:
\enumerate{
\item model: the summary of the linear model, calculated with the function "lm(model.concentrations~model.absorbance)"
\item Pr: probability of the model
\item R.sq: Adjusted R squared of the model
\item formula: Formula of the model
\item predicted.concentrations: Vector containing the concentrations calculated with the formula obtained from the model
}
}
\description{
The function creates the linear model, based on the absorbance of samples with known concentration. If added, the function can also calculate the concentrations of further samples based on the absorbance
}
\keyword{absorbance}
\keyword{lm,}
