% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TtSpectra.R
\name{Tt.Spectra}
\alias{Tt.Spectra}
\title{Plot spectra for EDS analysis}
\usage{
Tt.Spectra(energy, counts, elements.list, y.lim = c(0, max(counts)),
  x.lim = c(0, max(energy)))
}
\arguments{
\item{energy}{vector containing the values for the x axe. Typically in KeV for EDS analysis}

\item{counts}{vector containing the values for the y axe}

\item{elements.list}{a vector containing the simbol of the elements to be displayed in the graphic}

\item{y.lim}{vector containing limits for y axe}

\item{x.lim}{vector containing limits for x axe}
}
\value{
spectra plot
}
\description{
The function uses basic R plot for plotting spectra. However, if elements.list is provided, it will provide labels for the elements, based on the energy (KeV) table of the elements for EDS analysis.
}
\examples{
Tt.spectra(spectrum$Energy, spectrum$Counts,elements.list=c('Pb','Zn','P'), y.lim=c(0,4000), x.lim=c(0,11))
}
\keyword{EDS}
\keyword{spectra}
