% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ttspectra.R
\name{Tt.spectra2}
\alias{Tt.spectra2}
\title{Plot spectra for EDS analysis}
\usage{
Tt.spectra2(energy, counts, elements.list = NULL, lab.size = 1,
  y.lim = c(0, max(counts, na.rm = T)), x.lim = c(0, max(energy, na.rm
  = T)), as.ggplot = TRUE)
}
\arguments{
\item{energy}{vector containing the values for the x axe. Typically in KeV for EDS analysis}

\item{counts}{vector containing the values for the y axe}

\item{elements.list}{a vector containing the simbol of the elements to be displayed in the graphic}

\item{lab.size}{size of the labels. Note that this size is considerably different for ggplot2 and basic R}

\item{y.lim}{vector containing limits for y axe}

\item{x.lim}{vector containing limits for x axe}

\item{as.ggplot}{creates the plot using 'ggplot2' and 'scales' packages. If false, it will use base R packages}
}
\value{
spectra plot
}
\description{
The function uses basic R plot for plotting spectra. However, if elements.list is provided, it will provide labels for the elements, based on the energy (KeV) table of the elements for EDS analysis.
}
\examples{
Tt.spectra(spectrum$Energy, spectrum$Counts,elements.list=c('Pb','Zn','P'), y.lim=c(0,4000), x.lim=c(0,11), as.ggplot=F) # Simple plot with personilised limits
Tt.spectra(spectrum$Energy, spectrum$Counts,elements.list=c('Ca'),as.ggplot=F) # When there is overlap, as Ca and Sb, is best to add second element manually
text(3.5, 500, 'Prueba', cex = 0.8)
Tt.spectra(spectrum$Energy, spectrum$Counts,elements.list=c('Ca')) # However this has been fixed using 'ggplot2'
}
\keyword{EDS}
\keyword{spectra}
