% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TtTable.R
\name{Tt.table}
\alias{Tt.table}
\title{Writes a table with mean +/- SD, all as character}
\usage{
Tt.table(frames.list, row.list, col.list, decimals = 2,
  separator = "±")
}
\arguments{
\item{frames.list}{list containing the data frame with the values. The data frame "stats" obtained from function "Tt.aov" can be used directly here. Otherwise, make sure to name "x.mean" and "x.sd" your columns with mean and standard deviation values, respectively.}

\item{row.list}{vector containing the names of the rows (must match with row length)}

\item{col.list}{vector containing the names of the columns (must match with columns length)}

\item{decimals}{amount of decimals after point. Default is 2}

\item{separator}{symbol to separate values. Default is "\u00B1"}
}
\value{
a data frame with mean and standard deviation values, set together in a single row as character. Useful to export.
}
\description{
Writes a table with mean +/- SD, all as character
}
\keyword{export}
\keyword{mean}
\keyword{standard-deviation}
\keyword{table}
