% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ttbarplot.R
\name{Tt.barplot}
\alias{Tt.barplot}
\title{Bar chart with standard error lines using ggplot}
\usage{
Tt.barplot(aov.output, T1, T2 = NULL, e.lines = "SD", dif = NULL,
  dif.h = 0.5)
}
\arguments{
\item{aov.output}{is a data frame containing the statistics information with the following data: mean values, treatments, error lines, and if desired, letters representing statistical difference.
The object \emph{stats} obtained from the function \code{Tt.aov()} can be used directly.}

\item{T1}{name of the column containing the treatments}

\item{T2}{name of the column containing the treatments for 2-factors plot (optional)}

\item{e.lines}{name of the column containing the data of the error lines to be plotted}

\item{dif}{name of the column with the letters representing statistical differences}

\item{dif.h}{distance from the uper error line to plot the \code{dif} letters}
}
\value{
An object for ggplot graphics
}
\description{
Bar chart with standard error lines using ggplot
}
\examples{
Zn <- Tt.aov(compost$Pb,compost$Treatment,compost$Plant) # 2-way anova
plot <- Tt.barplot(Zn$stats, treat2=T)
plot + labs(title="Zn in plant tissue", x='Species', y='mg/Kg')+
scale_fill_brewer(palette=3)+
guides(fill=guide_legend(title='Treatment'))
}
\keyword{anova}
\keyword{barplot}
\keyword{ggplot}
