% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBarplot.R
\name{Tt.barplot}
\alias{Tt.barplot}
\title{Bar chart with standard error lines using ggplot}
\usage{
Tt.barplot(aov.output, treat2 = FALSE)
}
\arguments{
\item{aov.output}{is a data frame containing the statistics information with the following data. Column names must match with the described below:
\describe{
\item{Treatment(s)}{The vector with the independent variables. Should be stored in the first columns of the data frame (1 or 2). No name matching is necessary.}
\item{Mean}{Vector with mean values}
\item{SD}{vector with standard deviation}
\item{n}{vector with the sample sizes}
}
The object \emph{stats} obtained from the function \code{Tt.aov()} can be used directly, as it has the specified format.}

\item{treat2=FALSE}{plots 1-way anova. If a second independent variable exists, change to TRUE}
}
\value{
gBarPlo An object for ggplot graphics
}
\description{
Bar chart with standard error lines using ggplot
}
\examples{
2anova.Zn <- Tt.aov(compost$Pb,compost$Treatment,compost$Plant) # 2-way anova
plot <- Tt.barplot(2anova.Zn$stats, treat2=T)
plot + labs(title="Graphic Title", subtitle='Subtitle', x='labels in X', y='labels in Y') + scale_fill_brewer(palette=3, name='Treatments')
}
\keyword{anova}
\keyword{barplot}
\keyword{ggplot}
