% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ttppm.R
\name{Tt.ppm}
\alias{Tt.ppm}
\title{Calculation of ppm or ppb of a solid}
\usage{
Tt.ppm(solid, solution, concentration, mol = NULL, units = list(solid =
  "g", solution = "ml", concentration = "mg"),
  result.units = list(element = "mg", sample = "Kg"))
}
\arguments{
\item{solid}{mass of the solid processed in a solution}

\item{solution}{volume of the solution where the solid was processed}

\item{concentration}{concentration obtained after the lecture}

\item{mol}{if some units are expressed as mol, molar weght should be specified}

\item{units}{a list containing units of the solid, solution and concentration. See details}

\item{result.units}{mass units of element per mass units of sample}
}
\value{
returns the proportion of element contained in a solid, expressed in terms specified under "result.units"
}
\description{
The function calculates proportion of element contained in a solid, express as mass units of element per mass units of sample (g, ug, mg, Kg), after being processed through a solution, particularly useful after reading absorbance AA, ICP, etc.
}
\details{
explain concentration

explain units
}
\examples{
# 0.1014g of plant were processed in 7.5ml of water
# The laboratory results give 13.009 uM/L proline (molar weight 75.066)
# We want to express the results as mg of proline per Kg of plant
Tt.ppm(solid=0.1014,solution=7.5,concentration=13.009,mol=75.066,units=list(solid='g',solution='ml',concentration='uM'), result.units=list(element='mg',sample='Kg'))

# Soil was digested in aqua regia, and diluted at 25ml for ICP
soil <- c(0.266,0.271,0.277,0.243,0.257) # weight of samples
As <- c(0.054968685,0.056348685,0.090648685,0.057058685,0.070963685) # results from ICP
Pb <- c(0.837786373333334,0.916599706666667,0.886459706666667,0.86488304,0.981849706666667)
icp <- data.frame(As,Pb)
Tt.ppm(soil,25,icp)
}
